+++ src/backend/azure.rs
@@ -37,6 +37,7 @@ use crate::UrlExt;
 use crate::backend::StorageBackend;
 use crate::backend::Upload;
 use crate::backend::auth::azure::RequestSigner;
+use crate::backend::format_range_header;
 use crate::generator::Alphanumeric;
 use crate::streams::ByteStream;
 use crate::streams::TransferStream;
@@ -622,15 +623,23 @@ impl StorageBackend for AzureBlobStorageBackend {
         Ok(response)
     }
 
-    async fn get_at_offset(&self, url: Url, etag: &str, offset: u64) -> Result<Response> {
+    async fn get_range(
+        &self,
+        url: Url,
+        etag: &str,
+        start: u64,
+        exclusive_end: Option<u64>,
+    ) -> Result<Response> {
         debug_assert!(
             Self::is_supported_url(&self.config, &url),
             "{url} is not a supported Azure URL",
             url = url.as_str()
         );
 
+        let range = format_range_header(start, exclusive_end);
+
         debug!(
-            "sending GET request at offset {offset} for `{url}`",
+            "sending GET request with range `{range}` for `{url}`",
             url = url.display(),
         );
 
@@ -639,7 +648,7 @@ impl StorageBackend for AzureBlobStorageBackend {
             .get(url)
             .header(header::USER_AGENT, USER_AGENT)
             .header(header::DATE, Utc::now().to_http_date())
-            .header(header::RANGE, format!("bytes={offset}-"))
+            .header(header::RANGE, range)
             .header(header::IF_MATCH, etag)
             .header(AZURE_VERSION_HEADER, AZURE_STORAGE_VERSION)
             .build()?;