+++ src/backend/s3.rs
@@ -34,6 +34,7 @@ use crate::backend::StorageBackend;
 use crate::backend::Upload;
 use crate::backend::auth::s3::RequestSigner;
 use crate::backend::auth::s3::SignatureProvider;
+use crate::backend::format_range_header;
 use crate::sha256_hex_string;
 use crate::streams::ByteStream;
 use crate::streams::TransferStream;
@@ -700,15 +701,23 @@ impl StorageBackend for S3StorageBackend {
         Ok(response)
     }
 
-    async fn get_at_offset(&self, url: Url, etag: &str, offset: u64) -> Result<Response> {
+    async fn get_range(
+        &self,
+        url: Url,
+        etag: &str,
+        start: u64,
+        exclusive_end: Option<u64>,
+    ) -> Result<Response> {
         debug_assert!(
             Self::is_supported_url(&self.config, &url),
             "{url} is not a supported S3 URL",
             url = url.as_str()
         );
 
+        let range = format_range_header(start, exclusive_end);
+
         debug!(
-            "sending GET request at offset {offset} for `{url}`",
+            "sending GET request with range `{range}` for `{url}`",
             url = url.display(),
         );
 
@@ -720,7 +729,7 @@ impl StorageBackend for S3StorageBackend {
             .header(header::USER_AGENT, USER_AGENT)
             .header(AWS_DATE_HEADER, date.format("%Y%m%dT%H%M%SZ").to_string())
             .header(AWS_CONTENT_SHA256_HEADER, sha256_hex_string([]))
-            .header(header::RANGE, format!("bytes={offset}-"))
+            .header(header::RANGE, range)
             .header(header::IF_MATCH, etag)
             .build()?;