+++ src/config.rs
@@ -15,9 +15,12 @@ use base64::prelude::BASE64_STANDARD;
 use secrecy::SecretString;
 use serde::Deserialize;
 use sha2::Digest;
+use tokio::select;
 use tokio::task::spawn_blocking;
 use tokio_retry2::strategy::ExponentialFactorBackoff;
 use tokio_retry2::strategy::MaxInterval;
+use tokio_util::sync::CancellationToken;
+use tracing::info;
 
 /// The default number of retries for network operations.
 const DEFAULT_RETRIES: usize = 5;
@@ -43,12 +46,23 @@ pub enum HashAlgorithm {
 impl HashAlgorithm {
     /// Calculates a `Content-Digest` header value for the contents of the file
     /// at the given path.
-    pub async fn calculate_content_digest(&self, path: &Path) -> crate::Result<Option<String>> {
+    ///
+    /// Returns `None` if content digests are not being calculated for uploads.
+    pub async fn calculate_content_digest(
+        &self,
+        path: &Path,
+        cancel: &CancellationToken,
+    ) -> crate::Result<Option<String>> {
+        info!(
+            "calculating content digest for file `{path}`",
+            path = path.display()
+        );
+
         match self {
             Self::None => Ok(None),
             Self::Sha256 => {
                 let path = path.to_path_buf();
-                spawn_blocking(move || {
+                let fut = spawn_blocking(move || {
                     let mut hasher = sha2::Sha256::new();
                     let mut reader = BufReader::new(File::open(path)?);
                     std::io::copy(&mut reader, &mut hasher)?;
@@ -57,13 +71,16 @@ impl HashAlgorithm {
                         "sha-256=:{encoded}:",
                         encoded = BASE64_STANDARD.encode(digest)
                     )))
-                })
-                .await
-                .expect("failed to join task")
+                });
+
+                select! {
+                    _ = cancel.cancelled() => Err(crate::Error::Canceled),
+                    r = fut => r.expect("failed to join task")
+                }
             }
             Self::Blake3 => {
                 let path = path.to_path_buf();
-                spawn_blocking(move || {
+                let fut = spawn_blocking(move || {
                     let mut hasher = blake3::Hasher::new();
                     hasher.update_mmap_rayon(&path)?;
                     let digest = hasher.finalize();
@@ -71,9 +88,12 @@ impl HashAlgorithm {
                         "blake3=:{encoded}:",
                         encoded = BASE64_STANDARD.encode(digest.as_bytes())
                     )))
-                })
-                .await
-                .expect("failed to join task")
+                });
+
+                select! {
+                    _ = cancel.cancelled() => Err(crate::Error::Canceled),
+                    r = fut => r.expect("failed to join task")
+                }
             }
         }
     }
@@ -425,14 +445,11 @@ impl ConfigBuilder {
 
     /// Sets the parallelism supported for uploads and downloads.
     ///
-    /// For uploads, this is the number of blocks that may be concurrently
-    /// transferred for a single file.
-    ///
-    /// For downloads, this is the number of files that may be concurrently
-    /// downloaded.
+    /// This is the maximum number of concurrent transfer streams that may be
+    /// attempted at one time.
     ///
     /// Defaults to the host's available parallelism (or 1 if it cannot be
-    /// determined).
+    /// determined) multiplied by 2.
     pub fn with_parallelism(mut self, parallelism: usize) -> Self {
         self.parallelism = Some(parallelism);
         self
@@ -440,16 +457,13 @@ impl ConfigBuilder {
 
     /// Sets the parallelism supported for uploads and downloads.
     ///
-    /// For uploads, this is the number of blocks that may be concurrently
-    /// transferred for a single file.
-    ///
-    /// For downloads, this is the number of files that may be concurrently
-    /// downloaded.
+    /// This is the maximum number of concurrent transfer streams that may be
+    /// attempted at one time.
     ///
     /// If `None`, the default parallelism is used.
     ///
     /// Defaults to the host's available parallelism (or 1 if it cannot be
-    /// determined).
+    /// determined) multiplied by 2.
     pub fn with_maybe_parallelism(mut self, parallelism: Option<usize>) -> Self {
         self.parallelism = parallelism;
         self
@@ -605,17 +619,14 @@ impl Config {
 
     /// Gets the parallelism supported for uploads and downloads.
     ///
-    /// For uploads, this is the number of blocks that may be concurrently
-    /// transferred for a single file.
-    ///
-    /// For downloads, this is the number of files that may be concurrently
-    /// downloaded.
+    /// This is the number of blocks that may be concurrently transferred for a
+    /// single file.
     ///
     /// Defaults to the host's available parallelism (or 1 if it cannot be
-    /// determined).
+    /// determined) and multiplied by 2.
     pub fn parallelism(&self) -> usize {
         self.parallelism
-            .unwrap_or_else(|| available_parallelism().map(NonZero::get).unwrap_or(1))
+            .unwrap_or_else(|| available_parallelism().map(NonZero::get).unwrap_or(1) * 2)
     }
 
     /// Sets the parallelism supported for uploads and downloads.