+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+#### Added
+
+* Added support for downloading files using multiple streams in parallel;
+  requires not using a download cache as the download cache does not support
+  ranged requests ([#34](https://github.com/stjude-rust-labs/cloud-copy/pull/34)).
+
 ## 0.7.0 - 02-04-2026
 
 #### Fixed